+++ Source/Core/Utilities/NovusFileUtils.pas
@@ -21,6 +21,7 @@ Type
     class function IsFileReadonly(fName: string): boolean;
     class function MoveDir(aFromDirectory, aToDirectory: String): boolean;
     class function CopyDir(aFromDirectory, aToDirectory: String): boolean;
+    class function ExtractName(aFullFileName: String): String;
     /// <summary>
     /// Extracts the extension part of a full file name without "."
     /// </summary>
@@ -344,6 +345,11 @@ begin
   lsTmpFilename := ExtractFilename(aFolder);
 
   Result := ((lsTmpFilename = '') and (lsTmpFolder <> ''));
+end;
+
+class function TNovusFileUtils.ExtractName(aFullFileName: String): String;
+begin
+  Result := stringreplace(extractfilename(aFullFileName), ExtractFileExt(aFullFileName), '', [rfIgnoreCase]);
 end;
 
 end.