+++ Source/Core/Utilities/NovusNumUtils.pas
@@ -3,7 +3,8 @@ unit NovusNumUtils;
 
 interface
 
-uses Windows, sysutils, NovusUtilities;
+uses Windows, sysutils, NovusUtilities, System.StrUtils,
+     System.Classes;
 
 Type
   TNovusNumUtils = class(TNovusUtilities)
@@ -19,6 +20,12 @@ Type
     class function BinToInt64(BinStr: string): Int64;
     class function BinToUInt64(BinStr: string): UInt64;
     class function BinToUInt8(BinStr: string): UInt8;
+    class function BinToUInt16(BinStr: string): UInt16;
+    class function StrToUInt16(aStr: String): UInt16;
+    class function CreateWord(const aHiByte, aLoByte: byte): word;
+    class function HexStrToBytes(aHexStr: String; aLittle_Endian: boolean = False): tBytes;
+    class function ReverseBytesOrder(aBytes: TBytes): tBytes;
+
   end;
 
 implementation
@@ -104,6 +111,31 @@ class function TNovusNumUtils.BinToUInt8(BinStr: string): UInt8;
 var
   i: byte;
   RetVar: UInt8;
+begin
+  Result := 0;
+  if Trim(BinStr) = '' then Exit;
+
+  BinStr := UpperCase(BinStr);
+  if BinStr[length(BinStr)] = 'B' then
+    Delete(BinStr, length(BinStr), 1);
+  RetVar := 0;
+  for i := 1 to length(BinStr) do
+  begin
+    if not(BinStr[i] in ['0', '1']) then
+    begin
+      RetVar := 0;
+      Break;
+    end;
+    RetVar := (RetVar shl 1) + (byte(BinStr[i]) and 1);
+  end;
+
+  result := RetVar;
+end;
+
+class function TNovusNumUtils.BinToUInt16(BinStr: string): UInt16;
+var
+  i: byte;
+  RetVar: UInt16;
 begin
   Result := 0;
   if Trim(BinStr) = '' then Exit;
@@ -213,6 +245,19 @@ begin
 
 end;
 
+class function TNovusNumUtils.StrToUInt16(aStr: String): UInt16;
+begin
+  Result := 0;
+  if Trim(aStr) = '' then Exit;
+
+  Try
+    Result := StrToUInt(aStr);
+  Except
+    Result := 0;
+  End;
+
+end;
+
 
 
 class function TNovusNumUtils.HexToUInt16(HexStr: String): Uint16;
@@ -274,6 +319,44 @@ begin
 
   result := RetVar;
 end;
+
+
+class function TNovusNumUtils.CreateWord(const aHiByte, aLoByte: byte): word;
+var
+  WR: WordRec;
+begin
+  WR.Hi := aHiByte;
+  WR.Lo := aLoByte;
+  result := word(WR);
+end;
+
+class function TNovusNumUtils.HexStrToBytes(aHexStr: String; aLittle_Endian: boolean): tBytes;
+Var
+  TmpBytes: TBytes;
+begin
+  SetLength(TmpBytes, Length(aHexStr) div 2);
+  HexToBin(PChar(aHexStr), TmpBytes[0], Length(TmpBytes));
+
+  if aLittle_Endian then
+    TmpBytes :=  ReverseBytesOrder(TmpBytes);
+
+  Result := TmpBytes;
+end;
+
+class function TNovusNumUtils.ReverseBytesOrder(aBytes: TBytes): tBytes;
+var
+  TmpBytes: TBytes;
+  i, h: integer;
+begin
+  h := High(aBytes);
+  SetLength(TmpBytes, Length(aBytes));
+  for i := 0 to h do
+    TmpBytes[i] := aBytes[h - i];
+  Result := TmpBytes;
+end;
+
+//LittleEndian
+